+++ tests/photons_transport_tests/test_script.py
@@ -1,6 +1,6 @@
 # coding: spec
 
-from photons_transport.target.script import ScriptRunner, ScriptRunnerIterator
+from photons_transport.target.script import ScriptRunner
 
 from photons_app.errors import PhotonsAppError, BadRunWithResults
 from photons_app.test_helpers import AsyncTestCase
@@ -22,67 +22,6 @@ describe AsyncTestCase, "ScriptRunner":
         target = mock.Mock(name="target")
 
         runner = ScriptRunner(script, target)
-
-        self.assertIs(runner.script, script)
-        self.assertIs(runner.target, target)
-
-    describe "run_with":
-        async before_each:
-            self.called = []
-
-            self.afr = mock.Mock(name="afr")
-            self.run_with_result = mock.Mock(name="run_with_result")
-
-            class FakeScript(object):
-                async def run_with(s, *args, **kwargs):
-                    self.called.append(("run_with", args, kwargs))
-                    return self.run_with_result
-
-            class FakeTarget(object):
-                async def args_for_run(s, *args, **kwargs):
-                    self.called.append(("args_for_run", args, kwargs))
-                    return self.afr
-
-                async def close_args_for_run(s, *args, **kwargs):
-                    self.called.append(("close_args_for_run", args, kwargs))
-
-            self.script = FakeScript()
-            self.target = FakeTarget()
-            self.runner = ScriptRunner(self.script, self.target)
-
-        async it "calls run_with on the script":
-            self.assertEqual(self.called, [])
-
-            a = mock.Mock(name="a")
-            reference = mock.Mock(name="reference")
-            args_for_run = mock.NonCallableMock(name="args_for_run", spec=[])
-
-            self.assertIs(await self.runner.run_with(reference, args_for_run=args_for_run, b=a), self.run_with_result)
-            self.assertEqual(self.called
-                , [ ("run_with", (reference, args_for_run), {"b": a})
-                  ]
-                )
-
-        async it "creates and closes the afr if none provided":
-            self.assertEqual(self.called, [])
-
-            a = mock.Mock(name="a")
-            reference = mock.Mock(name="reference")
-
-            self.assertIs(await self.runner.run_with([reference], b=a), self.run_with_result)
-            self.assertEqual(self.called
-                , [ ("args_for_run", (), {})
-                  , ("run_with", ([reference], self.afr), {"b": a})
-                  , ("close_args_for_run", (self.afr, ), {})
-                  ]
-                )
-
-describe AsyncTestCase, "ScriptRunnerIterator":
-    async it "takes in script and target":
-        script = mock.Mock(name="script")
-        target = mock.Mock(name="target")
-
-        runner = ScriptRunnerIterator(script, target)
 
         self.assertIs(runner.script, script)
         self.assertIs(runner.target, target)
@@ -110,7 +49,7 @@ describe AsyncTestCase, "ScriptRunnerIterator":
 
         self.script = FakeScript()
         self.target = FakeTarget()
-        self.runner = ScriptRunnerIterator(self.script, self.target)
+        self.runner = ScriptRunner(self.script, self.target)
 
     describe "run_with":
         async it "calls run_with on the script":
@@ -252,7 +191,7 @@ describe AsyncTestCase, "ScriptRunnerIterator":
                     yield self.res1
                     raise error1
 
-            runner = ScriptRunnerIterator(FakeScript(), self.target)
+            runner = ScriptRunner(FakeScript(), self.target)
 
             self.assertEqual(self.called, [])